Move useObservable tests to react-hooks-testing-library
